+++ packages/bridge-controller/src/index.ts
@@ -157,6 +157,7 @@ export {
   selectExchangeRateByChainIdAndAddress,
   selectIsQuoteExpired,
   selectBridgeFeatureFlags,
+  selectMinimumBalanceForRentExemptionInSOL,
 } from './selectors';
 
 export { DEFAULT_FEATURE_FLAG_CONFIG } from './constants/bridge';